Remove isDestinationAddressExist from Near transaction metadata
Refactored TransactionLoadMetadata and related mapping/extensions to remove the isDestinationAddressExist parameter from the Near case. Updated usages in GatewayService and NearSigner to match the new signature. This simplifies the Near transaction metadata structure and removes unused logic.

+++ Packages/Signer/Sources/Chains/CosmosSigner.swift
@@ -47,7 +47,7 @@ public struct CosmosSigner: Signable {
             $0.accountNumber = UInt64(try input.metadata.getAccountNumber())
             $0.chainID = try input.metadata.getChainId()
             $0.memo = memo.valueOrEmpty
-            $0.sequence = UInt64(try input.metadata.getSequence())
+            $0.sequence = try input.metadata.getSequence()
             $0.messages = messages
             $0.fee = fee
             $0.privateKey = privateKey
