+++ Packages/Signer/Sources/Chains/PolkadotSigner.swift
@@ -13,7 +13,7 @@ public struct PolkadotSigner: Signable {
         let input = try PolkadotSigningInput.with {
             $0.genesisHash = data.genesisHash
             $0.blockHash = data.blockHash
-            $0.nonce = UInt64(try input.metadata.getSequence())
+            $0.nonce = try input.metadata.getSequence()
             $0.specVersion = data.specVersion
             $0.network = CoinType.polkadot.ss58Prefix
             $0.transactionVersion = data.transactionVersion
